Refactor CsvExporterTests.cs to remove unnecessary variables and update method call

+++ generate-covertura.cmd
@@ -1,6 +1,6 @@
  rem echo off
  echo
  rem - remove older test results
-rmdir /s /q CodeLineCounter.Tests\TestResults
+if exist rmdir /s /q CodeLineCounter.Tests\TestResults
  rem - collect code coverage
 dotnet test --collect:"XPlat Code Coverage;Format=json,lcov,cobertura"  --results-directory CodeLineCounter.Tests\TestResults --logger trx;LogFileName=testresults.trx
